fix(main): 메인 페이지 조회 - 주요 카테고리

+++ src/main/java/com/example/wanted_preonboarding_challenge_backend_31/shared/dto/category/CategoryFeaturedDto.java
@@ -5,6 +5,6 @@ public record CategoryFeaturedDto(
         String name,
         String slug,
         String imageUrl,
-        int productCount
+        Long productCount
 ) {
 }
